CS60 HW2: Java tests separated (only tests for one function per file)
Before, all the tests for Hw2pr8 were in one Java file. However, if
one function had issues in a student submission, this caused the whole
test file to fail and the student got no points. Now, they will just
get zero points for that function.
